+++ core-effects/src/main/java/dev/lec/effectapp/effects/EffectRegistry.kt
@@ -13,6 +13,7 @@ object EffectRegistry {
         SwirlEffect(),
         SpinEffect(),
         WaveEffect(),
+        TilesEffect(),
         PinchBulgeEffect(),
         GhostTrailEffect(),
         ColorInvertEffect(),